+++ Mapping/ClassMetadata.php
@@ -119,14 +119,53 @@ class ClassMetadata extends GenericMetadata implements ClassMetadataInterface
         }
     }
 
+    public function __serialize(): array
+    {
+        if (self::class === (new \ReflectionMethod($this, '__sleep'))->class) {
+            return [
+                'constraints' => $this->constraints,
+                'constraintsByGroup' => $this->constraintsByGroup,
+                'traversalStrategy' => $this->traversalStrategy,
+                'autoMappingStrategy' => $this->autoMappingStrategy,
+                'getters' => $this->getters,
+                'groupSequence' => $this->groupSequence,
+                'groupSequenceProvider' => $this->groupSequenceProvider,
+                'groupProvider' => $this->groupProvider,
+                'members' => $this->members,
+                'name' => $this->name,
+                'properties' => $this->properties,
+                'defaultGroup' => $this->defaultGroup,
+            ];
+        }
+
+        trigger_deprecation('symfony/validator', '7.4', 'Implementing "%s::__sleep()" is deprecated, use "__serialize()" instead.', get_debug_type($this));
+
+        $data = [];
+        foreach ($this->__sleep() as $key) {
+            try {
+                if (($r = new \ReflectionProperty($this, $key))->isInitialized($this)) {
+                    $data[$key] = $r->getValue($this);
+                }
+            } catch (\ReflectionException) {
+                $data[$key] = $this->$key;
+            }
+        }
+
+        return $data;
+    }
+
+    /**
+     * @deprecated since Symfony 7.4, will be removed in 8.0
+     */
     public function __sleep(): array
     {
-        $parentProperties = parent::__sleep();
+        trigger_deprecation('symfony/validator', '7.4', 'Calling "%s::__sleep()" is deprecated, use "__serialize()" instead.', get_debug_type($this));
 
-        // Don't store the cascading strategy. Classes never cascade.
-        unset($parentProperties[array_search('cascadingStrategy', $parentProperties)]);
-
-        return array_merge($parentProperties, [
+        return [
+            'constraints',
+            'constraintsByGroup',
+            'traversalStrategy',
+            'autoMappingStrategy',
             'getters',
             'groupSequence',
             'groupSequenceProvider',
@@ -135,7 +174,7 @@ class ClassMetadata extends GenericMetadata implements ClassMetadataInterface
             'name',
             'properties',
             'defaultGroup',
-        ]);
+        ];
     }
 
     public function getClassName(): string